+++ server/src/main/java/org/allaymc/server/entity/component/player/EntityPlayerContainerHolderComponentImpl.java
@@ -180,6 +180,7 @@ public class EntityPlayerContainerHolderComponentImpl extends EntityContainerHol
             }
 
             if (entityArrow.isPickUpDisabled()) {
+                entityArrow.remove();
                 continue;
             }